@@ -17,27 +17,39 @@ namespace EpicLoot_UnityLib
         public void OnEnable()
         {
             _focusedListIndex = 0;
-            for (var index = 0; index < Lists.Count; index++)
+            for (int index = 0; index < Lists.Count; index++)
             {
+                if (Lists[index] == null)
+                {
+                    continue;
+                }
                 Lists[index].GiveFocus(index == _focusedListIndex, 0);
             }
+
             RefreshHints();
         }
 
         public void Update()
         {
             if (Lists.Count == 0)
+            {
                 return;
+            }
 
-            var currentList = Lists[_focusedListIndex];
-            var loopCount = 0;
-            var itemCount = currentList.GetItemCount();
+            MultiSelectItemList currentList = Lists[_focusedListIndex];
+            int loopCount = 0;
+            int itemCount = currentList.GetItemCount();
             while (itemCount == 0)
             {
                 currentList.GiveFocus(false, 0);
 
                 _focusedListIndex = (_focusedListIndex + 1) % Lists.Count;
                 currentList = Lists[_focusedListIndex];
+                if (currentList == null)
+                {
+                    continue;
+                }
+
                 itemCount = currentList.GetItemCount();
                 if (currentList.GetItemCount() > 0)
                 {
@@ -45,14 +57,17 @@ namespace EpicLoot_UnityLib
                     RefreshHints();
                     break;
                 }
+
                 loopCount++;
                 if (loopCount >= Lists.Count)
+                {
                     return;
+                }
             }
 
             if (ZInput.IsGamepadActive())
             {
-                var newFocusedIndex = _focusedListIndex;
+                int newFocusedIndex = _focusedListIndex;
                 if (ZInput.GetButtonDown("JoyTabLeft"))
                 {
                     newFocusedIndex = Mathf.Max(_focusedListIndex - 1, 0);
@@ -66,33 +81,39 @@ namespace EpicLoot_UnityLib
 
                 if (newFocusedIndex != _focusedListIndex)
                 {
-                    var offset = newFocusedIndex - _focusedListIndex;
+                    int offset = newFocusedIndex - _focusedListIndex;
                     if (Lists[newFocusedIndex].GetItemCount() == 0)
+                    {
                         newFocusedIndex = (newFocusedIndex + offset + Lists.Count) % Lists.Count;
+                    }
                     if (Lists[newFocusedIndex].GetItemCount() == 0)
+                    {
                         newFocusedIndex = _focusedListIndex;
+                    }
                 }
 
                 FocusList(newFocusedIndex);
             }
                 
             if (_gamepadWasEnabled != ZInput.IsGamepadActive())
+            {
                 RefreshHints();
+            }
 
             _gamepadWasEnabled = ZInput.IsGamepadActive();
         }
 
         public void FocusList(int newFocusedIndex)
         {
-            var list = Lists[_focusedListIndex];
-            var currentFocusElement = list.GetFocusedElement();
-            var currentFocusIndex = currentFocusElement != null ? currentFocusElement.transform.GetSiblingIndex() : -1;
+            MultiSelectItemList list = Lists[_focusedListIndex];
+            MultiSelectItemListElement currentFocusElement = list.GetFocusedElement();
+            int currentFocusIndex = currentFocusElement != null ? currentFocusElement.transform.GetSiblingIndex() : -1;
             if (newFocusedIndex != _focusedListIndex && newFocusedIndex >= 0 && newFocusedIndex < Lists.Count)
             {
                 _focusedListIndex = newFocusedIndex;
-                for (var index = 0; index < Lists.Count; index++)
+                for (int index = 0; index < Lists.Count; index++)
                 {
-                    var isGrid = Lists[index].IsGrid();
+                    bool isGrid = Lists[index].IsGrid();
                     Lists[index].GiveFocus(index == _focusedListIndex, isGrid ? 0 : currentFocusIndex);
                 }
 
@@ -105,16 +126,18 @@ namespace EpicLoot_UnityLib
             if (!isActiveAndEnabled || !ZInput.IsGamepadActive() || Lists.Count == 0)
                 return;
 
-            var focusedList = Lists[_focusedListIndex];
-            foreach (var hint in SortHints)
+            MultiSelectItemList focusedList = Lists[_focusedListIndex];
+            foreach (GameObject hint in SortHints)
             {
-                hint.SetActive(focusedList.Sortable && focusedList.SortByDropdown != null && focusedList.SortByDropdown.isActiveAndEnabled);
+                hint.SetActive(focusedList.Sortable && focusedList.SortByDropdown != null &&
+                    focusedList.SortByDropdown.isActiveAndEnabled);
             }
-            foreach (var hint in SelectAllHints)
+            foreach (GameObject hint in SelectAllHints)
             {
-                hint.SetActive(focusedList.Multiselect && focusedList.SelectAllToggle != null && focusedList.SelectAllToggle.isActiveAndEnabled);
+                hint.SetActive(focusedList.Multiselect && focusedList.SelectAllToggle != null &&
+                    focusedList.SelectAllToggle.isActiveAndEnabled);
             }
-            foreach (var hint in SelectHints)
+            foreach (GameObject hint in SelectHints)
             {
                 hint.SetActive(!focusedList.ReadOnly && focusedList.GetFocusedElement() != null);
             }
