@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+using System;
+using System.Collections.Generic;
 
 namespace EpicLoot.Abilities
 {
@@ -6,10 +7,12 @@ namespace EpicLoot.Abilities
     {
         public static AbilityConfig Config;
         public static readonly Dictionary<string, AbilityDefinition> Abilities = new Dictionary<string, AbilityDefinition>();
+        public static event Action OnSetupAbilityDefinitions;
 
         public static void Initialize(AbilityConfig config)
         {
             Config = config;
+            OnSetupAbilityDefinitions?.Invoke();
 
             Abilities.Clear();
             foreach (var def in Config.Abilities)
@@ -24,6 +27,11 @@ namespace EpicLoot.Abilities
                         $"Please fix your configuration.");
                 }
             }
+        }
+
+        public static AbilityConfig GetCFG()
+        {
+            return Config;
         }
 
         public static bool TryGetAbilityDef(string abilityID, out AbilityDefinition abilityDef)
