@@ -1,4 +1,5 @@
-﻿using EpicLoot.Adventure.Feature;
+using EpicLoot.Adventure.Feature;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
@@ -15,10 +16,14 @@ namespace EpicLoot.Adventure
         public static TreasureMapsAdventureFeature TreasureMaps;
         public static BountiesAdventureFeature Bounties;
         public static int CheatNumberOfBounties = -1;
-
+        #nullable enable
+        public static event Action? OnSetupAdventureData;
+        #nullable disable
         public static void Initialize(AdventureDataConfig config)
         {
             Config = config;
+            
+            OnSetupAdventureData?.Invoke();
 
             SecretStash = new SecretStashAdventureFeature();
             Gamble = new GambleAdventureFeature();
@@ -26,6 +31,12 @@ namespace EpicLoot.Adventure
             Bounties = new BountiesAdventureFeature();
 
             Config.TreasureMap.UpdateBiomeList();
+            EpicLoot.Log($"Updated/setup Adventure Data");
+        }
+
+        public static AdventureDataConfig GetCFG()
+        {
+            return Config;
         }
 
         public static void UpdateAventureData(AdventureDataConfig config)
@@ -33,6 +44,7 @@ namespace EpicLoot.Adventure
             Config = config;
 
             Config.TreasureMap.UpdateBiomeList();
+            EpicLoot.Log($"Updated Adventure Data");
         }
 
         public static Sprite GetTrophyIconForMonster(string monsterID, bool isGold)
