@@ -1,4 +1,4 @@
-﻿using EpicLoot_UnityLib;
+using EpicLoot_UnityLib;
 using System.Linq;
 using TMPro;
 using UnityEngine;
@@ -10,14 +10,16 @@ namespace EpicLoot
     {
         public static Button ReplaceButton(Button button, bool icon = false, bool keepListeners = false)
         {
-            var newButton = Auga.API.MediumButton_Create(button.transform.parent, button.name, string.Empty);
-            return ReplaceButtonInternal(newButton, button, icon, keepListeners);
+            //var newButton = Auga.API.MediumButton_Create(button.transform.parent, button.name, string.Empty);
+            // Note first argument should be the new button, if Auga is fixed.
+            return ReplaceButtonInternal(button, button, icon, keepListeners);
         }
 
         public static Button ReplaceButtonFancy(Button button, bool icon = false, bool keepListeners = false)
         {
-            var newButton = Auga.API.FancyButton_Create(button.transform.parent, button.name, string.Empty);
-            return ReplaceButtonInternal(newButton, button, icon, keepListeners);
+            //var newButton = Auga.API.FancyButton_Create(button.transform.parent, button.name, string.Empty);
+            // Note first argument should be the new button, if Auga is fixed.
+            return ReplaceButtonInternal(button, button, icon, keepListeners);
         }
 
         private static Button ReplaceButtonInternal(Button newButton, Button button, bool icon, bool keepListeners)
@@ -64,59 +66,59 @@ namespace EpicLoot
 
         public static void MakeSimpleTooltip(GameObject obj)
         {
-            Auga.API.Tooltip_MakeSimpleTooltip(obj);
+            //Auga.API.Tooltip_MakeSimpleTooltip(obj);
         }
 
         public static void ReplaceBackground(GameObject obj, bool withCornerDecoration)
         {
-            var image = obj.GetComponent<Image>();
-            Object.Destroy(image);
-            var backgroundPanel = Auga.API.Panel_Create(obj.transform, Vector2.one, "AugaBackground", withCornerDecoration);
-            var rt = (RectTransform)backgroundPanel.transform;
-            rt.SetSiblingIndex(0);
-            rt.anchorMin = Vector2.zero;
-            rt.anchorMax = Vector2.one;
-            rt.pivot = new Vector2(0.5f, 0.5f);
-            rt.sizeDelta = new Vector2(40, 40);
+            //var image = obj.GetComponent<Image>();
+            //Object.Destroy(image);
+            //var backgroundPanel = Auga.API.Panel_Create(obj.transform, Vector2.one, "AugaBackground", withCornerDecoration);
+            //var rt = (RectTransform)backgroundPanel.transform;
+            //rt.SetSiblingIndex(0);
+            //rt.anchorMin = Vector2.zero;
+            //rt.anchorMax = Vector2.one;
+            //rt.pivot = new Vector2(0.5f, 0.5f);
+            //rt.sizeDelta = new Vector2(40, 40);
         }
 
         public static void FixItemBG(GameObject obj)
         {
-            var magicBG = obj.transform.Find("MagicBG");
-            if (magicBG != null)
-            {
-                var image = magicBG.GetComponent<Image>();
-                if (image != null)
-                {
-                    image.sprite = EpicLoot.GetMagicItemBgSprite();
-                }
-            }
+            //var magicBG = obj.transform.Find("MagicBG");
+            //if (magicBG != null)
+            //{
+            //    var image = magicBG.GetComponent<Image>();
+            //    if (image != null)
+            //    {
+            //        image.sprite = EpicLoot.GetMagicItemBgSprite();
+            //    }
+            //}
 
-            var replacedBG = false;
-            var itemBG = obj.transform.Find("ItemBG");
-            var baseImage = itemBG != null ? itemBG.GetComponent<Image>() : null;
-            if (baseImage == null)
-                baseImage = obj.GetComponent<Image>();
+            //var replacedBG = false;
+            //var itemBG = obj.transform.Find("ItemBG");
+            //var baseImage = itemBG != null ? itemBG.GetComponent<Image>() : null;
+            //if (baseImage == null)
+            //    baseImage = obj.GetComponent<Image>();
 
-            if (baseImage != null)
-            {
-                baseImage.sprite = Auga.API.GetItemBackgroundSprite();
-                baseImage.color = new Color(0, 0, 0, 0.5f);
-                replacedBG = true;
-            }
+            //if (baseImage != null)
+            //{
+            //    baseImage.sprite = Auga.API.GetItemBackgroundSprite();
+            //    baseImage.color = new Color(0, 0, 0, 0.5f);
+            //    replacedBG = true;
+            //}
 
-            if (!replacedBG)
-            {
-                var icon = obj.transform.Find("Icon");
-                if (icon != null)
-                {
-                    var iconBG = Object.Instantiate(icon, icon.parent);
-                    iconBG.SetSiblingIndex(2);
-                    var image = iconBG.GetComponent<Image>();
-                    image.sprite = Auga.API.GetItemBackgroundSprite();
-                    image.color = new Color(0, 0, 0, 0.5f);
-                }
-            }
+            //if (!replacedBG)
+            //{
+            //    var icon = obj.transform.Find("Icon");
+            //    if (icon != null)
+            //    {
+            //        var iconBG = Object.Instantiate(icon, icon.parent);
+            //        iconBG.SetSiblingIndex(2);
+            //        var image = iconBG.GetComponent<Image>();
+            //        image.sprite = Auga.API.GetItemBackgroundSprite();
+            //        image.color = new Color(0, 0, 0, 0.5f);
+            //    }
+            //}
         }
 
         public static void FixListElementColors(GameObject obj)
@@ -124,12 +126,12 @@ namespace EpicLoot
             var selected = obj.transform.Find("Selected");
             if (selected != null)
             {
-                var image = selected.GetComponent<Image>();
-                if (image != null)
-                {
-                    ColorUtility.TryParseHtmlString(Auga.API.Blue, out var color);
-                    image.color = color;
-                }
+                //var image = selected.GetComponent<Image>();
+                //if (image != null)
+                //{
+                //    ColorUtility.TryParseHtmlString(Auga.API.Blue, out var color);
+                //    image.color = color;
+                //}
             }
 
             var background = obj.transform.Find("Background");
@@ -154,25 +156,25 @@ namespace EpicLoot
             var texts = obj.GetComponentsInChildren<TMP_Text>(true);
             foreach (var text in texts)
             {
-                if (text.font.name == "Norsebold")
-                {
-                    //text.font = Auga.API.GetBoldFont();
-                    ColorUtility.TryParseHtmlString(Auga.API.Brown1, out var color);
-                    text.color = color;
-                    text.text = Localization.instance.Localize(text.text).ToUpperInvariant();
-                }
-                else
-                {
-                    //text.font = Auga.API.GetSemiBoldFont();
-                }
+                //if (text.font.name == "Norsebold")
+                //{
+                //    //text.font = Auga.API.GetBoldFont();
+                //    ColorUtility.TryParseHtmlString(Auga.API.Brown1, out var color);
+                //    text.color = color;
+                //    text.text = Localization.instance.Localize(text.text).ToUpperInvariant();
+                //}
+                //else
+                //{
+                //    //text.font = Auga.API.GetSemiBoldFont();
+                //}
 
-                if (text.name == "Count" || text.name == "RewardLabel" || text.name.EndsWith("Count"))
-                {
-                    ColorUtility.TryParseHtmlString(Auga.API.BrightGold, out var color);
-                    text.color = color;
-                }
+                //if (text.name == "Count" || text.name == "RewardLabel" || text.name.EndsWith("Count"))
+                //{
+                //    ColorUtility.TryParseHtmlString(Auga.API.BrightGold, out var color);
+                //    text.color = color;
+                //}
 
-                text.text = text.text.Replace("<color=yellow>", $"<color={Auga.API.BrightGold}>");
+                //text.text = text.text.Replace("<color=yellow>", $"<color={Auga.API.BrightGold}>");
             }
         }
 
