@@ -154,19 +154,14 @@ namespace EpicLoot
     /// The following class is largely taken from Azumatt's Tooltip Expansion mod:
     /// https://github.com/AzumattDev/TooltipExpansion/blob/main/CodeNShit/Monos/TooltipSizeAdjuster.cs
     /// </summary>
-    public class ScrollWheelHandler : MonoBehaviour
-    {
+    public class ScrollWheelHandler : MonoBehaviour {
         private ScrollRect _scrollRect = null;
 
-        public void Awake()
-        {
+        public void Awake() {
             _scrollRect = GetComponent<ScrollRect>();
-            if (_scrollRect == null)
-            {
+            if (_scrollRect == null) {
                 EpicLoot.LogWarning("ScrollWheelHandler: No ScrollRect found on " + gameObject.name);
-            }
-            else
-            {
+            } else {
                 _scrollRect.verticalNormalizedPosition = 1f;
             }
         }
@@ -175,17 +170,11 @@ namespace EpicLoot
         {
             // Only process if this object is active.
             if (!gameObject.activeInHierarchy || _scrollRect == null)
-            {
                 return;
-            }
 
             // Get scroll wheel input regardless of pointer location.
             float scrollDelta = Input.GetAxis("Mouse ScrollWheel");
-            if (!(Mathf.Abs(scrollDelta) > float.Epsilon))
-            {
-                return;
-            }
-
+            if (!(Mathf.Abs(scrollDelta) > float.Epsilon)) return;
             // Adjust the vertical scroll position.
             float newScrollPosition = _scrollRect.verticalNormalizedPosition + scrollDelta * 0.7f;
             _scrollRect.verticalNormalizedPosition = Mathf.Clamp01(newScrollPosition);
