@@ -1,5 +1,6 @@
 ﻿using EpicLoot.Config;
 using HarmonyLib;
+using System.IO;
 using UnityEngine;
 using UnityEngine.UI;
 
@@ -7,24 +8,28 @@ namespace EpicLoot
 {
     public class WelcomeMessage : MonoBehaviour
     {
-        public const string PlayerPrefKey = "el-wm";
-        public const int VersionNumber = 1;
-
-        public static bool PlayerHasSeenMessage()
-        {
-            return PlayerPrefs.GetInt(PlayerPrefKey, 0) >= VersionNumber;
-        }
-
-        public static void SetPlayerHasSeenMessage()
-        {
-            PlayerPrefs.SetInt(PlayerPrefKey, VersionNumber);
+        public static void SetPlayerHasSeenMessage() {
+            ELConfig.AlwaysShowWelcomeMessage.Value = false;
         }
 
         public void Awake()
         {
+            var titleText = transform.Find("Title")?.GetComponent<Text>();
+            titleText.text = Localization.instance.Localize(titleText.text) + $"{EpicLoot.Version}";
+
+            var contentText = transform.Find("Content")?.GetComponent<Text>();
+            contentText.text = Localization.instance.Localize(contentText.text);
+
             var discordButton = transform.Find("DiscordButton")?.GetComponent<Button>();
             var patchNotesButton = transform.Find("PatchNotesButton")?.GetComponent<Button>();
             var closeButton = transform.Find("CloseButton")?.GetComponent<Button>();
+
+            var overhaulMinimalButton = transform.Find("overhaul_minimal")?.GetComponent<Button>();
+            overhaulMinimalButton?.onClick.AddListener(SetOverhaulMinimalAndClick);
+            var overhaulBalancedButton = transform.Find("overhaul_balanced")?.GetComponent<Button>();
+            overhaulBalancedButton?.onClick.AddListener(SetOverhaulBalancedAndClick);
+            var overhaulLegendaryButton = transform.Find("overhaul_legendary")?.GetComponent<Button>();
+            overhaulLegendaryButton?.onClick.AddListener(SetOverhaulLegendaryAndClick);
 
             if (EpicLoot.HasAuga)
             {
@@ -63,6 +68,39 @@ namespace EpicLoot
             SetPlayerHasSeenMessage();
             Destroy(gameObject);
         }
+
+        public void SetOverhaulBalancedAndClick()
+        {
+            ELConfig.BalanceConfigurationType.Value = "balanced";
+            ELConfig.ItemsUnidentifiedDropRatio.Value = 0.8f;
+            ELConfig.ItemsToMaterialsDropRatio.Value = 0.95f;
+            OnOverhaulButtomClick();
+            Close();
+        }
+
+        public void SetOverhaulMinimalAndClick()
+        {
+            ELConfig.BalanceConfigurationType.Value = "minimal";
+            ELConfig.ItemsToMaterialsDropRatio.Value = 1.0f;
+            OnOverhaulButtomClick();
+            Close();
+        }
+
+        public void SetOverhaulLegendaryAndClick()
+        {
+            ELConfig.BalanceConfigurationType.Value = "legendary";
+            ELConfig.ItemsUnidentifiedDropRatio.Value = 0.2f;
+            ELConfig.ItemsToMaterialsDropRatio.Value = 0.1f;
+            OnOverhaulButtomClick();
+            Close();
+        }
+
+        public void OnOverhaulButtomClick() {
+            string basecfglocation = Path.Combine(ELConfig.GetOverhaulDirectoryPath(), "magiceffects.json");
+            var overhaulfiledata = EpicLoot.ReadEmbeddedResourceFile(ELConfig.GetDefaultEmbeddedFileLocation("magiceffects.json"));
+            File.WriteAllText(basecfglocation, overhaulfiledata);
+        }
+
     }
 
     [HarmonyPatch(typeof(FejdStartup), nameof(FejdStartup.Start))]
@@ -70,11 +108,13 @@ namespace EpicLoot
     {
         public static void Postfix(FejdStartup __instance)
         {
-            if (ELConfig.AlwaysShowWelcomeMessage.Value || !WelcomeMessage.PlayerHasSeenMessage())
+            if (ELConfig.AlwaysShowWelcomeMessage.Value)
             {
                 var welcomeMessage = Object.Instantiate(EpicLoot.Assets.WelcomMessagePrefab, __instance.transform, false);
                 welcomeMessage.name = "WelcomeMessage";
                 welcomeMessage.AddComponent<WelcomeMessage>();
+                welcomeMessage.GetComponent<WelcomeMessage>();
+                ELConfig.AlwaysShowWelcomeMessage.Value = false;
             }
         }
     }
