@@ -38,7 +38,7 @@ public class ProjectileMover : MonoBehaviour
     {
 		if (speed != 0)
         {
-            rb.velocity = transform.forward * speed;
+            rb.linearVelocity = transform.forward * speed;
             //transform.position += transform.forward * (speed * Time.deltaTime);         
         }
 	}
